move stuff to globals
